+++ src/mappers/mod.rs
@@ -1,3 +1,4 @@
+pub mod follow_mapper;
 pub mod post_mapper;
 pub mod task_mapper;
 pub mod user_mapper;