+++ src/routes/mod.rs
@@ -1,3 +1,4 @@
+pub mod follow_route;
 pub mod post_route;
 pub mod task_route;
 pub mod user_route;