+++ src/services/impl_service/mod.rs
@@ -1,3 +1,4 @@
+pub mod follow_impl;
 pub mod post_impl;
 pub mod task_impl;
 pub mod user_impl;