+++ src/services/mod.rs
@@ -1,3 +1,4 @@
+pub mod follow_service;
 pub mod impl_service;
 pub mod post_service;
 pub mod task_service;